cloudreactor-python-api-client/pyproject.toml
[tool.poetry]
name = "cloudreactor-python-api-client"
version = "0.2.0"
description = "A client library for accessing CloudReactor API"

authors = [
    "Jeff Tsay <jeff@cloudreactor.io>"
]
license = "BSD 2-Clause"
readme = "README.md"
repository = "https://github.com/CloudReactor/cloudreactor-python-api-client/cloudreactor-python-api-client"
documentation = "https://cloudreactor-procwrapper.readthedocs.io"
homepage = "https://cloudreactor.io"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
]

packages = [
    {include = "cloudreactor_python_api_client"},
]
include = ["CHANGELOG.md", "cloudreactor_python_api_client/py.typed"]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/CloudReactor/cloudreactor-python-api-client/issues"

[tool.poetry.dependencies]
python = "^3.7"
httpx = ">=0.15.4,<0.23.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.20.0"
flake8 = "^3.8.4"
isort = "^5.6.4"
pytest = "^6.2.1"
pytest-cov = "^2.10.1"
pytest-env = "^0.6.2"
tox = ">=3.25.0"
mypy = ">=0.942"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target_version = ['py37', 'py38', 'py39']
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
  )/
)
'''

[tool.isort]
line_length = 120
profile = "black"